django/requirements.txt
Django>=3.0.3
djangorestframework>=3.11.0
django-filter>=2.2.0
gunicorn>=19.9.0
mysqlclient>=1.4.4

